requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.16.2
pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.0
protobuf==3.20.3
mysql-connector-python==8.1.0